+++ documentation/changelog.rst
@@ -15,7 +15,7 @@ New features
 
 Bugfixes
 -----------
-
+* Choose better forecasting horizons when weather data is posted [see `PR #131 <http://www.github.com/SeitaBV/flexmeasures/pull/131>`_]
 
 Infrastructure / Support
 ----------------------